Add files via upload
Salary of an employee is calculated where the user enter its name, its dateofbirth and it then calculates its age. If it is greater than or equal to 18 he/she is an eligible employee. It then asks for its basic salary and then we calculate its gross salary by adding da,ta,hra and deducting pf from it also if its salary is greater than 10000 then 1000 is deducted as professinal tax. Also we calculate its retirement year ie on which year he/she will get retired.